requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
apturl==0.5.2
async-timeout==4.0.2
attrs==23.1.0
ayatana-settings==21.1.28
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cajarename==21.11.24
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
deja-dup-caja==0.0.9
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
fasteners==0.14.1
folder-color-caja==0.0.86
folder-color-common==0.0.86
frozenlist==1.4.0
future==0.18.2
gpg===1.16.0-unknown
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
Magnus==1.0.3
Mako==1.1.3
MarkupSafe==2.0.1
mate-hud==22.4.4
mate-tweak==22.4.8
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
psutil==5.9.0
ptyprocess==0.7.0
pulsemixer==1.5.1
py-cord==2.4.1
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-xlib==0.29
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==1.0.0